Add files via upload
Pencil ads were previously a non-standard size, did not preserve aspect ratio, and disappeared on mobile.

This commit updates pencil ads to standard (798x90) size and places them in the middle of the screen. In mobile, these ads are replaced by 468x60 ads. A new sidebar is registered to allow Boston Hassle to upload a different image for this aspect ratio.